+++ sdk/elastic/arm-elastic/samples/v3-beta/javascript/vmCollectionUpdateSample.js
@@ -1,0 +1,24 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+const { MicrosoftElastic } = require("@azure/arm-elastic");
+const { DefaultAzureCredential } = require("@azure/identity");
+
+/**
+ * This sample demonstrates how to update the VM details that will be monitored by the Elastic monitor resource, ensuring optimal observability and performance.
+ *
+ * @summary update the VM details that will be monitored by the Elastic monitor resource, ensuring optimal observability and performance.
+ * x-ms-original-file: 2025-06-01/VMCollection_Update.json
+ */
+async function vmCollectionUpdate() {
+  const credential = new DefaultAzureCredential();
+  const subscriptionId = "00000000-0000-0000-0000-000000000000";
+  const client = new MicrosoftElastic(credential, subscriptionId);
+  await client.vmCollection.update("myResourceGroup", "myMonitor");
+}
+
+async function main() {
+  await vmCollectionUpdate();
+}
+
+main().catch(console.error);