+++ sdk/elastic/arm-elastic/src/classic/index.ts
@@ -1,0 +1,28 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export type { AllTrafficFiltersOperations } from "./allTrafficFilters/index.js";
+export type { AssociateTrafficFilterOperations } from "./associateTrafficFilter/index.js";
+export type { BillingInfoOperations } from "./billingInfo/index.js";
+export type { ConnectedPartnerResourcesOperations } from "./connectedPartnerResources/index.js";
+export type { CreateAndAssociateIPFilterOperations } from "./createAndAssociateIPFilter/index.js";
+export type { CreateAndAssociatePLFilterOperations } from "./createAndAssociatePLFilter/index.js";
+export type { DeploymentInfoOperations } from "./deploymentInfo/index.js";
+export type { DetachAndDeleteTrafficFilterOperations } from "./detachAndDeleteTrafficFilter/index.js";
+export type { DetachTrafficFilterOperations } from "./detachTrafficFilter/index.js";
+export type { ElasticVersionsOperations } from "./elasticVersions/index.js";
+export type { ExternalUserOperations } from "./externalUser/index.js";
+export type { ListAssociatedTrafficFiltersOperations } from "./listAssociatedTrafficFilters/index.js";
+export type { MonitorOperations } from "./monitor/index.js";
+export type { MonitoredResourcesOperations } from "./monitoredResources/index.js";
+export type { MonitoredSubscriptionsOperations } from "./monitoredSubscriptions/index.js";
+export type { MonitorsOperations } from "./monitors/index.js";
+export type { OpenAIOperations } from "./openAI/index.js";
+export type { OperationsOperations } from "./operations/index.js";
+export type { OrganizationsOperations } from "./organizations/index.js";
+export type { TagRulesOperations } from "./tagRules/index.js";
+export type { TrafficFiltersOperations } from "./trafficFilters/index.js";
+export type { UpgradableVersionsOperations } from "./upgradableVersions/index.js";
+export type { VMCollectionOperations } from "./vmCollection/index.js";
+export type { VMHostOperations } from "./vmHost/index.js";
+export type { VMIngestionOperations } from "./vmIngestion/index.js";